pyproto: split into new feature
