+++ pyo3-macros-backend/src/pyclass.rs
@@ -511,11 +511,11 @@ fn unit_variants_as_descriptors<'a>(
         .map(|var| gen_py_const(&cls_type, &variant_to_attribute(var)));
 
     quote! {
-        impl _pyo3::class::impl_::PyClassDescriptors<#cls>
-            for _pyo3::class::impl_::PyClassImplCollector<#cls>
+        impl _pyo3::impl_::pyclass::PyClassDescriptors<#cls>
+            for _pyo3::impl_::pyclass::PyClassImplCollector<#cls>
         {
-            fn py_class_descriptors(self) -> &'static [_pyo3::class::methods::PyMethodDefType] {
-                static METHODS: &[_pyo3::class::methods::PyMethodDefType] = &[#(#py_methods),*];
+            fn py_class_descriptors(self) -> &'static [_pyo3::impl_::pymethods::PyMethodDefType] {
+                static METHODS: &[_pyo3::impl_::pymethods::PyMethodDefType] = &[#(#py_methods),*];
                 METHODS
             }
         }
@@ -574,11 +574,11 @@ fn impl_descriptors(
         .collect::<syn::Result<_>>()?;
 
     Ok(quote! {
-        impl _pyo3::class::impl_::PyClassDescriptors<#cls>
-            for _pyo3::class::impl_::PyClassImplCollector<#cls>
+        impl _pyo3::impl_::pyclass::PyClassDescriptors<#cls>
+            for _pyo3::impl_::pyclass::PyClassImplCollector<#cls>
         {
-            fn py_class_descriptors(self) -> &'static [_pyo3::class::methods::PyMethodDefType] {
-                static METHODS: &[_pyo3::class::methods::PyMethodDefType] = &[#(#py_methods),*];
+            fn py_class_descriptors(self) -> &'static [_pyo3::impl_::pymethods::PyMethodDefType] {
+                static METHODS: &[_pyo3::impl_::pymethods::PyMethodDefType] = &[#(#py_methods),*];
                 METHODS
             }
         }
@@ -676,7 +676,7 @@ impl<'a> PyClassImplsBuilder<'a> {
         };
 
         let base_nativetype = if attr.has_extends {
-            quote! { <Self::BaseType as _pyo3::class::impl_::PyClassBaseType>::BaseNativeType }
+            quote! { <Self::BaseType as _pyo3::impl_::pyclass::PyClassBaseType>::BaseNativeType }
         } else {
             quote! { _pyo3::PyAny }
         };
@@ -749,13 +749,13 @@ impl<'a> PyClassImplsBuilder<'a> {
         };
 
         let thread_checker = if self.attr.has_unsendable {
-            quote! { _pyo3::class::impl_::ThreadCheckerImpl<#cls> }
+            quote! { _pyo3::impl_::pyclass::ThreadCheckerImpl<#cls> }
         } else if self.attr.has_extends {
             quote! {
-                _pyo3::class::impl_::ThreadCheckerInherited<#cls, <#cls as _pyo3::class::impl_::PyClassImpl>::BaseType>
+                _pyo3::impl_::pyclass::ThreadCheckerInherited<#cls, <#cls as _pyo3::impl_::pyclass::PyClassImpl>::BaseType>
             }
         } else {
-            quote! { _pyo3::class::impl_::ThreadCheckerStub<#cls> }
+            quote! { _pyo3::impl_::pyclass::ThreadCheckerStub<#cls> }
         };
 
         let (for_each_py_method, methods_protos, inventory, inventory_class) = match self
@@ -775,13 +775,13 @@ impl<'a> PyClassImplsBuilder<'a> {
                 );
                 (
                     quote! {
-                        for inventory in _pyo3::inventory::iter::<<Self as _pyo3::class::impl_::PyClassImpl>::Inventory>() {
-                            visitor(_pyo3::class::impl_::PyClassInventory::methods(inventory));
+                        for inventory in _pyo3::inventory::iter::<<Self as _pyo3::impl_::pyclass::PyClassImpl>::Inventory>() {
+                            visitor(_pyo3::impl_::pyclass::PyClassInventory::methods(inventory));
                         }
                     },
                     quote! {
-                        for inventory in _pyo3::inventory::iter::<<Self as _pyo3::class::impl_::PyClassImpl>::Inventory>() {
-                            visitor(_pyo3::class::impl_::PyClassInventory::slots(inventory));
+                        for inventory in _pyo3::inventory::iter::<<Self as _pyo3::impl_::pyclass::PyClassImpl>::Inventory>() {
+                            visitor(_pyo3::impl_::pyclass::PyClassInventory::slots(inventory));
                         }
                     },
                     Some(quote! { type Inventory = #inventory_class_name; }),
@@ -789,8 +789,37 @@ impl<'a> PyClassImplsBuilder<'a> {
                 )
             }
         };
+
+        let pyproto_method_visitors = if cfg!(feature = "pyproto") {
+            Some(quote! {
+                visitor(collector.object_protocol_methods());
+                visitor(collector.async_protocol_methods());
+                visitor(collector.descr_protocol_methods());
+                visitor(collector.mapping_protocol_methods());
+                visitor(collector.number_protocol_methods());
+            })
+        } else {
+            None
+        };
+
+        let pyproto_proto_visitors = if cfg!(feature = "pyproto") {
+            Some(quote! {
+                visitor(collector.object_protocol_slots());
+                visitor(collector.number_protocol_slots());
+                visitor(collector.iter_protocol_slots());
+                visitor(collector.gc_protocol_slots());
+                visitor(collector.descr_protocol_slots());
+                visitor(collector.mapping_protocol_slots());
+                visitor(collector.sequence_protocol_slots());
+                visitor(collector.async_protocol_slots());
+                visitor(collector.buffer_protocol_slots());
+            })
+        } else {
+            None
+        };
+
         quote! {
-            impl _pyo3::class::impl_::PyClassImpl for #cls {
+            impl _pyo3::impl_::pyclass::PyClassImpl for #cls {
                 const DOC: &'static str = #doc;
                 const IS_GC: bool = #is_gc;
                 const IS_BASETYPE: bool = #is_basetype;
@@ -802,48 +831,36 @@ impl<'a> PyClassImplsBuilder<'a> {
                 #inventory
 
                 fn for_each_method_def(visitor: &mut dyn ::std::ops::FnMut(&[_pyo3::class::PyMethodDefType])) {
-                    use _pyo3::class::impl_::*;
+                    use _pyo3::impl_::pyclass::*;
                     let collector = PyClassImplCollector::<Self>::new();
                     #for_each_py_method;
                     visitor(collector.py_class_descriptors());
-                    visitor(collector.object_protocol_methods());
-                    visitor(collector.async_protocol_methods());
-                    visitor(collector.descr_protocol_methods());
-                    visitor(collector.mapping_protocol_methods());
-                    visitor(collector.number_protocol_methods());
+                    #pyproto_method_visitors
                 }
                 fn get_new() -> ::std::option::Option<_pyo3::ffi::newfunc> {
-                    use _pyo3::class::impl_::*;
+                    use _pyo3::impl_::pyclass::*;
                     let collector = PyClassImplCollector::<Self>::new();
                     collector.new_impl()
                 }
                 fn get_alloc() -> ::std::option::Option<_pyo3::ffi::allocfunc> {
-                    use _pyo3::class::impl_::*;
+                    use _pyo3::impl_::pyclass::*;
                     let collector = PyClassImplCollector::<Self>::new();
                     collector.alloc_impl()
                 }
                 fn get_free() -> ::std::option::Option<_pyo3::ffi::freefunc> {
-                    use _pyo3::class::impl_::*;
+                    use _pyo3::impl_::pyclass::*;
                     let collector = PyClassImplCollector::<Self>::new();
                     collector.free_impl()
                 }
 
                 fn for_each_proto_slot(visitor: &mut dyn ::std::ops::FnMut(&[_pyo3::ffi::PyType_Slot])) {
                     // Implementation which uses dtolnay specialization to load all slots.
-                    use _pyo3::class::impl_::*;
+                    use _pyo3::impl_::pyclass::*;
                     let collector = PyClassImplCollector::<Self>::new();
                     // This depends on Python implementation detail;
                     // an old slot entry will be overriden by newer ones.
                     visitor(collector.py_class_default_slots());
-                    visitor(collector.object_protocol_slots());
-                    visitor(collector.number_protocol_slots());
-                    visitor(collector.iter_protocol_slots());
-                    visitor(collector.gc_protocol_slots());
-                    visitor(collector.descr_protocol_slots());
-                    visitor(collector.mapping_protocol_slots());
-                    visitor(collector.sequence_protocol_slots());
-                    visitor(collector.async_protocol_slots());
-                    visitor(collector.buffer_protocol_slots());
+                    #pyproto_proto_visitors
                     #methods_protos
                 }
 
@@ -861,7 +878,7 @@ impl<'a> PyClassImplsBuilder<'a> {
 
         self.attr.freelist.as_ref().map_or(quote!{}, |freelist| {
             quote! {
-                impl _pyo3::class::impl_::PyClassWithFreeList for #cls {
+                impl _pyo3::impl_::pyclass::PyClassWithFreeList for #cls {
                     #[inline]
                     fn get_free_list(_py: _pyo3::Python<'_>) -> &mut _pyo3::impl_::freelist::FreeList<*mut _pyo3::ffi::PyObject> {
                         static mut FREELIST: *mut _pyo3::impl_::freelist::FreeList<*mut _pyo3::ffi::PyObject> = 0 as *mut _;
@@ -875,17 +892,17 @@ impl<'a> PyClassImplsBuilder<'a> {
                     }
                 }
 
-                impl _pyo3::class::impl_::PyClassAllocImpl<#cls> for _pyo3::class::impl_::PyClassImplCollector<#cls> {
+                impl _pyo3::impl_::pyclass::PyClassAllocImpl<#cls> for _pyo3::impl_::pyclass::PyClassImplCollector<#cls> {
                     #[inline]
                     fn alloc_impl(self) -> ::std::option::Option<_pyo3::ffi::allocfunc> {
-                        ::std::option::Option::Some(_pyo3::class::impl_::alloc_with_freelist::<#cls>)
+                        ::std::option::Option::Some(_pyo3::impl_::pyclass::alloc_with_freelist::<#cls>)
                     }
                 }
 
-                impl _pyo3::class::impl_::PyClassFreeImpl<#cls> for _pyo3::class::impl_::PyClassImplCollector<#cls> {
+                impl _pyo3::impl_::pyclass::PyClassFreeImpl<#cls> for _pyo3::impl_::pyclass::PyClassImplCollector<#cls> {
                     #[inline]
                     fn free_impl(self) -> ::std::option::Option<_pyo3::ffi::freefunc> {
-                        ::std::option::Option::Some(_pyo3::class::impl_::free_with_freelist::<#cls>)
+                        ::std::option::Option::Some(_pyo3::impl_::pyclass::free_with_freelist::<#cls>)
                     }
                 }
             }
@@ -928,7 +945,7 @@ fn define_inventory_class(inventory_class_name: &syn::Ident) -> TokenStream {
             }
         }
 
-        impl _pyo3::class::impl_::PyClassInventory for #inventory_class_name {
+        impl _pyo3::impl_::pyclass::PyClassInventory for #inventory_class_name {
             fn methods(&'static self) -> &'static [_pyo3::class::PyMethodDefType] {
                 self.methods
             }
