+++ pyo3-macros/Cargo.toml
@@ -16,6 +16,8 @@ proc-macro = true
 [features]
 multiple-pymethods = []
 
+pyproto = ["pyo3-macros-backend/pyproto"]
+
 [dependencies]
 proc-macro2 = { version = "1", default-features = false }
 quote = "1"
