+++ tests/test_arithmetics_protos.rs
@@ -1,5 +1,6 @@
 #![allow(deprecated)] // for deprecated protocol methods
 #![cfg(feature = "macros")]
+#![cfg(feature = "pyproto")]
 
 use pyo3::class::basic::CompareOp;
 use pyo3::class::*;
