+++ tests/test_mapping_pyproto.rs
@@ -1,4 +1,5 @@
 #![cfg(feature = "macros")]
+#![cfg(feature = "pyproto")]
 
 use std::collections::HashMap;
 
@@ -7,6 +8,7 @@ use pyo3::prelude::*;
 use pyo3::py_run;
 use pyo3::types::IntoPyDict;
 use pyo3::types::PyList;
+use pyo3::PyMappingProtocol;
 
 mod common;
 
@@ -32,7 +34,10 @@ impl Mapping {
             })
         }
     }
+}
 
+#[pyproto]
+impl PyMappingProtocol for Mapping {
     fn __len__(&self) -> usize {
         self.index.len()
     }
