+++ tests/ui/not_send.stderr
@@ -6,9 +6,9 @@ error[E0277]: `*mut pyo3::Python<'static>` cannot be shared between threads safe
     |
     = help: within `pyo3::Python<'_>`, the trait `Sync` is not implemented for `*mut pyo3::Python<'static>`
     = note: required because it appears within the type `PhantomData<*mut pyo3::Python<'static>>`
-    = note: required because it appears within the type `pyo3::impl_::not_send::NotSend`
-    = note: required because it appears within the type `(&GILGuard, pyo3::impl_::not_send::NotSend)`
-    = note: required because it appears within the type `PhantomData<(&GILGuard, pyo3::impl_::not_send::NotSend)>`
+    = note: required because it appears within the type `impl_::not_send::NotSend`
+    = note: required because it appears within the type `(&GILGuard, impl_::not_send::NotSend)`
+    = note: required because it appears within the type `PhantomData<(&GILGuard, impl_::not_send::NotSend)>`
     = note: required because it appears within the type `pyo3::Python<'_>`
     = note: required because of the requirements on the impl of `Send` for `&pyo3::Python<'_>`
     = note: required because it appears within the type `[closure@$DIR/tests/ui/not_send.rs:4:22: 4:38]`
